fix: don’t attributes to modular docs
* also use an asciidoctor fat jar
* clean up code a bit
