--- .build/binaries/splitter/JavaEWAH-1.1.7.jar
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:3ecf8b2c602314341f5a2ace171ed04fc86f2d4ddf762180656e9b71134ae68f
-size 166271
