--- .build/binaries/splitter/asciidoc-splitter-1.0-SNAPSHOT.jar
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:b3a5e26c92961282075fef7e9e87bdae081ad6fac12b0995c735be580f7d9d21
-size 50254
