--- .build/binaries/splitter/asciidoctorj-2.2.0.jar
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:7e92e9adf746803251acc5dc3cbc9c78d92b73c02a4aeb83b99fe401bbb6ec5b
-size 1701397
