--- .build/binaries/splitter/asciidoctorj-api-2.2.0.jar
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:d6b013a686902a021839f4368ed450f138b0c52265d09591c9686fee0188c7d7
-size 60079
