--- .build/binaries/splitter/backport9-1.3.jar
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:23315c0fd29bac16f7775a7cb1a47f7ace8e40387b7dc52de05797d76b2c1c36
-size 12924
