--- .build/binaries/splitter/bcpg-jdk15on-1.64.jar
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:10acaf221fc4e49d4a4067b02316271698e8742ef4b23cb5f2434a0e3502b7b4
-size 329081
