--- .build/binaries/splitter/bcpkix-jdk15on-1.64.jar
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:84669138b1d99143e2c009024f67824ab8d3edb9b05b7591f5ebfb020a4bda71
-size 878038
