--- .build/binaries/splitter/bcprov-jdk15on-1.64.jar
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:a4f463ce552b908a722fa198ef4892a226b3225e453f8df10d5c0a5bfe5db6b6
-size 4771726
