--- .build/binaries/splitter/dirgra-0.3.jar
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:9ee2b48dcdfe0f6261200b81852a3f8c02af702269457c82f485d36d6e078360
-size 16400
