--- .build/binaries/splitter/invokebinder-1.11.jar
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:96f5ffe37f13d4e7636c08467ec1e5c7af09a50c90baed50e2dc92e79faff7e9
-size 50026
