--- .build/binaries/splitter/javax.annotation-api-1.3.1.jar
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:bc1110630bb4290e798a533ca40a60517826c8804b79f91f8738d18ca425adc5
-size 26529
