--- .build/binaries/splitter/jboss-logmanager-2.1.14.Final.jar
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:5e47b4edc8b220d517a84322f7d25320d9c4be369307298fe133db5eddd67d80
-size 420926
