--- .build/binaries/splitter/jcodings-1.0.45.jar
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:dc9dcbca073690f6e3caa873c904293118f91d3105ad06a72af973b1cfdcf031
-size 1682629
