--- .build/binaries/splitter/jcommander-1.72.jar
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:e0de160b129b2414087e01fe845609cd55caec6820cfd4d0c90fabcc7bdb8c1e
-size 69254
