--- .build/binaries/splitter/jffi-1.2.22-native.jar
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:7883e6030b4ad53dc3be585a56ebcf9badfda52510ac9544f25423dc7e71439c
-size 789301
