--- .build/binaries/splitter/jffi-1.2.22.jar
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:f9ae61d6ed3e669cf58145c55845c39700023cf5c2c757633a3a1486444c8d6a
-size 152016
