--- .build/binaries/splitter/jnr-constants-0.9.14.jar
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:91c5d814a1c1e22c03377ee9ba3921867f15e65dea217b6b5ed381bde2c98df7
-size 654103
