--- .build/binaries/splitter/jnr-enxio-0.23.jar
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:f98896f39854651c5fb8cd5e6b83898dcb5de5df63d942b39e13af1aa52a29e1
-size 32037
