--- .build/binaries/splitter/jnr-netdb-1.1.6.jar
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:143e8bc45da8067aff83c88950802a7215fbf5a42b3cbf49c28ad1b6cc1e5959
-size 61647
