--- .build/binaries/splitter/jnr-posix-3.0.51.jar
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:7cfef38d9db97a29ade3adfec286c7967397a19d30beaf960a30994ffac34b9d
-size 254277
