--- .build/binaries/splitter/jnr-unixsocket-0.24.jar
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:c1738632edd8c9241b851cfea7fe3701d754c5d1da4c363d7948caa0a514f1a2
-size 47483
