--- .build/binaries/splitter/joni-2.1.30.jar
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:e8a1f991c3a323a21a06e0333f9de6c70933ec1b6691a12fd0e106de168aa608
-size 214805
