--- .build/binaries/splitter/jruby-9.2.9.0.jar
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:6a0a655568bcb63d88d83a03ae2738e5442130b697246280d28a030ec77769a5
-size 42049
