--- .build/binaries/splitter/jruby-core-9.2.9.0.jar
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:6e3f05f9cab80afcd58de9b34f36342ff4f37c3e39558619c4c9087b196dd51b
-size 10335228
