--- .build/binaries/splitter/jruby-stdlib-9.2.9.0.jar
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:2df33d039aafa2173abc2c37c15d4c8f2945530ab21238bcecbb7ece6b883c04
-size 11208976
