--- .build/binaries/splitter/jzlib-1.1.1.jar
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:5cb1e9f9cf0be011487545694ff0a178237c6bfcbb21c97865cdc52c60b9347a
-size 69578
