--- .build/binaries/splitter/nailgun-server-0.9.1.jar
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:4518faa6bf4bd26fccdc4d85e1625dc679381a08d56872d8ad12151dda9cef25
-size 32927
