--- .build/binaries/splitter/options-1.4.jar
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:7e1ac7036b75c9d66e5ae1931ca80724d3b49e93b8639d96ba5844b619984413
-size 14188
