--- .build/binaries/splitter/org.eclipse.jgit-5.7.0.202003110725-r.jar
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:ee13076c3f832048f1a0ca272b221d2b69f10919e3bfd5d1b1e5ff7cf661e417
-size 2854585
