--- .build/binaries/splitter/picocli-4.2.0.jar
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:d8cc74f00e6ae52e848539dfc7a05858dfe5cd19269491f8778efaa927665418
-size 364020
