--- .build/binaries/splitter/slf4j-api-1.7.2.jar
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:3bae789b401333b2a1d1603b7fa573e19908628191707203f6eb708cdee2c052
-size 26083
