--- .build/binaries/splitter/wildfly-common-1.5.1.Final.jar
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:39563538ca2f967aaffe9653f6722485ce43edb9aca94e50c01ffb6b2bbf64d0
-size 283915
